chore: Refactored repeated logic for Express SDK image operations
