+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "aem-genai-assistant",
-  "version": "1.0.4",
+  "version": "1.0.5",
   "private": true,
   "scripts": {
     "prepare": "husky install",
